Security Hardening
Read-protect /var/log/zabbix* and /var/run/zabbix* directories
so that only the zabbix agentd and server (plus root) can acces them.

Run zabbix server from a different user account as the agentd
so that if there is an agentd on the same system as the server
it can't see the database password in the server configuration
file.