Augmented find_trigger_ids and find_trigger_prototype_ids to take a expression in addition to the trigger name.  The intent of the find_trigger* functions is to find a specific trigger in the database, but trigger descriptions are not unique, only the combination of description and expression is unique.   These routines will also now filter out the instances of the "master" trigger definition as applied to hosts, as only the master is editable via the API.
Added several get_* routines which accept a structure of search
parameters, as defined by the Zabbix API, and return a list of
all matching Zabbix configuration entries.  This can be used to
perform an audit between what exists in Zabbix and what your
Chef config says should exist.  To do this it it best to set
a unique string in all your configuration entries, such as prepending
all Items with "Company-name: ".  It is then possible to fetch only
those Items and compare it to what you want to exist and create
resource blocks with "action :delete" to remove stale entries
via the Zabbix API.