redhat-lsb is not needed for the agent systemd script
This package pulls in everything and the kitchen sink and you really
don't want it on your server if you can avoid it. The RedHat SysVinit
script should also be rewritten to use the RedHat functions instead of
the LSB ones like the EPEL package does but I lack the time.